Fixed HMR & created Navbar prototype
Edited Dockerfile & docker-compose.yml in order to fix error where any updates would not show up.

Also added a Navbar component. This will need to be refactored to be more reusable - hard coding all values within navbar is not ideal. We will want to reuse this component later when implementing the sub-navbar