Created SubNavBar
Added a subnavbar with similar functionality to the Navbar. Navbar now passes an object called 'selected' which contains the chosen category and subCategory.

Also added a document listener so that clicking outside of Navbar/SubNavBar automatically minimizes the SubNavBar. This makes selecting a category without selecting a subCategory easy and will improve functionality.