created CardView
Created CardView module and adjusted functionality of card & cardList.

Now, when clicking on a card it routes you to CardView, which shows a more detailed version of the chosen card. It also contains a link to the website which opens in a different tabs.

Also lifted all card data to App.vue from Home.vue as it became an issue with data flow. It is easier to host all card data on App to ensure all components have access.