Services creation & no longer hardcoding URL
Created a folder for services and moved fetchAllCards function to it. This allows us to import it into App.vue - this makes for better structure.

Also created an .env file to hold our local host URL to prevent the URL from being hardcoded in the fetch functions.

Also created some error handling for the fetch function - given the server is not currently up it returns an HTLM/text type response when fetch is called - trying to parse this data type returns errors. This allows for clear explanation of where the issues may arise